+++ CHANGELOG.md
@@ -14,9 +14,10 @@
 - Renamed Activity sensor to Motion Activity
 - Renamed Audio sensor to Microphone
 - Renamed Air Pressure sensor to Barometer
+- Renamed Battery sensor to BatteryStatus
 - Renamed all sensor modules into sensors
 - Updated SensingKit-Android API (thanks to Ming-Jiun Huang @U3D3)
-- Updated Google Play Services to 15.0.1
+- Updated Google Play Services to 16.0.0
 
 ### 0.2.0 (July 30, 2015)
 - Added support for Bluetooth sensor